rename: ButtonProps interface types 폴더로 옮기기(#20)
