+++ src/components/ui/Button/Button.types.ts
@@ -1,5 +1,3 @@
-import type { ButtonHTMLAttributes, ReactNode } from 'react'
-
 export type ButtonStyle =
   | 'primary'
   | 'secondary'
@@ -9,13 +7,3 @@ export type ButtonStyle =
   | 'cancel' // wireframe의 닫기/취소 버튼용
 
 export type ButtonSize = 'small' | 'medium' | 'large'
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  btnStyle?: ButtonStyle
-  btnSize?: ButtonSize
-  btnIcon?: ReactNode
-  btnText?: string
-  iconOnly?: boolean
-  onClick?: (e: React.MouseEvent<HTMLButtonElement>) => void
-  disabled?: boolean
-}
