+++ src/components/ui/Button/index.ts
@@ -1,4 +1,4 @@
 export { default as Button } from './Button'
 export * from './Button.types'
 
-// import { Button, type ButtonProps, type ButtonStyle, type ButtonSize } from "@components/ui/Button"로 외부에서 사용 가능
+// import { Button, type ButtonStyle, type ButtonSize } from "@components/ui/Button"로 외부에서 사용 가능
